Add SuperGRAFX VPC library functions to HuCC.

+++ src/hucc/while.h
@@ -1,11 +1,11 @@
 #ifndef _WHILE_H
 #define _WHILE_H
 
-void addwhile (intptr_t *ptr);
+void addwhile (int *ptr);
 void delwhile (void);
-intptr_t *readwhile (void);
-intptr_t *findwhile (void);
-intptr_t *readswitch (void);
+int *readwhile (void);
+int *findwhile (void);
+int *readswitch (void);
 void addcase (int val);
 
 #endif
